// src/api/models/ProjectUpdate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ProjectUpdate = {
    profile_id?: (number | null);
    name?: (string | null);
    color?: (string | null);
    notes?: (string | null);
    archived?: (boolean | null);
};